pyproject.toml
[tool.poetry]
name = "projectd-drf"
version = "0.1.0"
description = ""
authors = ["shwwdnl <shwwdnl@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
setuptools = "^69.0.3"
djangorestframework = "^3.14.0"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
